requirements.txt
streamlit>=1.28.0
google-generativeai>=0.3.0
google-cloud-aiplatform>=1.38.0
flask>=3.0.0
flask-cors>=4.0.0
toml>=0.10.2
